+++ src/bunit.core/Rendering/FragmentContainer.cs
@@ -19,6 +19,10 @@ internal sealed class FragmentContainer : IComponent
 		{
 			renderHandle.Render(childContent);
 		}
+		else
+		{
+			renderHandle.Render(_ => {});
+		}
 
 		return Task.CompletedTask;
 	}